# api/openapi.yaml
openapi: 3.1.0
info:
  title: Onboarding Genérico — API de orquestación
  version: 0.1.0
  summary: Middleware B2B para la composición dinámica de flujos de onboarding e identidad digital.
  description: |
    API transaccional que consumen los sistemas **requirentes** (inquilinos B2B) para ejecutar
    flujos de onboarding compuestos dinámicamente a partir del Registro de Capacidades.

    ## Principios del contrato

    - **Ningún binario viaja por esta API.** Las imágenes y videos se suben directamente al
      almacenamiento de objetos mediante URLs prefirmadas obtenidas en `POST /v1/sessions/{sessionId}/artifacts`.
      El límite de 512 KB acumulados por ejecución de Cloud Workflows y el de 256 KiB por payload
      de Step Functions hacen inviable cualquier otro diseño.
    - **El inquilino se deriva del token, nunca del cuerpo ni de la ruta.** Un `tenantId` en el
      cuerpo de una petición se ignora; si contradice al del token, la petición se rechaza con `403`.
    - **Toda escritura es idempotente** mediante la cabecera `Idempotency-Key`.
    - **La decisión puede no ser del middleware.** En jurisdicciones donde la debida diligencia no
      es delegable —Bolivia, art. 32(II) del Instructivo UIF—, el flujo se configura con
      `decisionIssuer = SIGNALS_ONLY` y esta API devuelve señales, no un veredicto.
      Ver `docs/adr/0014-el-middleware-es-encargado-del-tratamiento.md`.

    ## Datos personales

...
    referencia y el resultado de la comparación.
  contact:
...
servers:
...
security:
  - bearerAuth: []
...
paths:
  /sessions:
    post:
      tags: [Sesiones]
      operationId: createSession
      summary: Inicia una sesión de onboarding
      description: |
        Resuelve la especificación de flujo aplicable a la combinación
        (inquilino, país, tipo de documento) consultando el Registro de Capacidades,
        la compila a un plan de ejecución y devuelve la lista de artefactos que el
        sujeto debe aportar.

        Si no existe una especificación activa para esa combinación, responde `422`
        con `code = FLOW_SPEC_NOT_FOUND`. **No se aplica ningún flujo por defecto**:
        adivinar el flujo sería adivinar el nivel de aseguramiento.
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/CorrelationId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateSessionRequest' }
            examples:
              cedulaBoliviana:
                summary: Cédula de identidad boliviana, flujo con señales sin veredicto
                value:
                  subjectRef: "cliente-9f3c2a"
                  countryCode: "BO"
                  documentType: "CI_BO"
                  flowName: "Standard-eKYC-Latam"
                  locale: "es-BO"
                  callbackUrl: "https://banco.example/webhooks/onboarding"
              pasaporteMexicano:
                summary: Pasaporte mexicano, flujo con veredicto automático
                value:
                  subjectRef: "usuario-4471"
                  countryCode: "MX"
                  documentType: "PASSPORT"
                  flowName: "Standard-eKYC-Latam"
      responses:
        '201':
          description: Sesión creada.
          headers:
            Location: { schema: { type: string, format: uri }, description: URL de la sesión creada. }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Session' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
...
  /sessions/{sessionId}/artifacts:
    parameters:
      - $ref: '#/components/parameters/SessionId'
    post:
      tags: [Artefactos]
      operationId: requestArtifactUpload
      summary: Solicita una URL prefirmada para cargar un artefacto
      description: |
        Devuelve una URL de carga directa al almacenamiento de objetos, con vencimiento corto.
        El binario **nunca** atraviesa esta API. Tras completar la carga, confirme con
        `POST /sessions/{sessionId}/artifacts/{artifactId}/complete` para que el plan avance.
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ArtifactUploadRequest' }
      responses:
        '201':
          description: URL de carga emitida.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ArtifactUploadTicket' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '413':
          description: El tamaño declarado excede el máximo admitido para ese tipo de artefacto.
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
...
  /sessions/{sessionId}/decision:
    parameters:
      - $ref: '#/components/parameters/SessionId'
    get:
      tags: [Decisión]
      operationId: getDecision
      summary: Obtiene el veredicto o las señales de la sesión
      description: |
        Cuando el flujo está configurado con `decisionIssuer = SIGNALS_ONLY`, el campo `outcome`
        vale `SIGNALS_ONLY` y la responsabilidad de decidir queda en el sistema requirente.
        Las señales y sus umbrales se devuelven íntegros para que la decisión del cliente sea auditable.
      responses:
        '200':
          description: Resultado disponible.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DecisionResult' }
        '202':
          description: La sesión aún no ha decidido. `Retry-After` indica cuándo reintentar.
          headers:
            Retry-After: { schema: { type: integer }, description: Segundos sugeridos. }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /reviews:
    get:
      tags: [Revisión humana]
      operationId: listPendingReviews
      summary: Lista las sesiones pendientes de revisión manual
      parameters:
        - name: state
          in: query
          schema: { type: string, enum: [PENDING_REVIEW, IN_REVIEW] }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
        - name: cursor
          in: query
          schema: { type: string }
      responses:
        '200':
          description: Página de casos.
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/ReviewCase' }
                  nextCursor: { type: string, nullable: true }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Token emitido por el proveedor de identidad federado del inquilino (Amazon Cognito o
        Identity Platform). Debe incluir el reclamo de inquilino. La autorización se resuelve en
        el núcleo, no en el gateway: Google Cloud API Gateway no admite autorizadores de código
        arbitrario, y mantener la lógica en el núcleo la hace portable entre ambas nubes.
  parameters:
    SessionId:
      name: sessionId
      in: path
      required: true
      schema: { type: string, minLength: 1, maxLength: 64 }
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: { type: string, format: uuid }
      description: |
        Identificador único de la operación. Repetir la misma petición con la misma clave
        devuelve el resultado original sin volver a ejecutar el flujo ni volver a facturar
        llamadas a proveedores externos.
    CorrelationId:
      name: X-Correlation-Id
      in: header
      required: false
      schema: { type: string, maxLength: 128 }
      description: Se propaga a trazas y registros. Si no se envía, el middleware lo genera.
  schemas:
    DocumentType:
      type: string
      enum: [PASSPORT, ID_CARD, RESIDENCE_PERMIT, DRIVING_LICENSE, INE_2019, INE_2021, CI_BO, CI_PY, UNKNOWN]
    CreateSessionRequest:
      type: object
      required: [subjectRef, countryCode, documentType]
      properties:
        subjectRef:
          type: string
          minLength: 1
          maxLength: 128
          description: |
            Referencia opaca al sujeto en el sistema del inquilino. **No use datos personales**
            como valor: ni nombre, ni número de documento, ni correo. Se persiste y aparece en
            el registro de auditoría.
        countryCode: { type: string, pattern: '^[A-Z]{2}$', description: País emisor del documento (ISO 3166-1 alfa-2). }
        documentType: { $ref: '#/components/schemas/DocumentType' }
        flowName:
          type: string
          maxLength: 64
          description: Nombre de la especificación de flujo. Si se omite, se resuelve por país y tipo de documento.
        locale: { type: string, example: es-BO }
        callbackUrl:
          type: string
          format: uri
          description: URL a la que se notificará el resultado. Debe ser HTTPS y estar en la lista permitida del inquilino.
        metadata:
          type: object
          maxProperties: 20
          additionalProperties: { type: string, maxLength: 256 }
          description: Metadatos opacos del inquilino. Prohibido incluir datos personales.
    Session:
      type: object
      required: [sessionId, state, flowName, countryCode, documentType, createdAt, steps, requiredArtifacts]
      properties:
        sessionId: { type: string }
        state: { $ref: '#/components/schemas/SessionState' }
        flowName: { type: string }
        flowVersion: { type: integer, description: Versión inmutable de la especificación con la que se compiló esta sesión. }
        countryCode: { type: string }
        documentType: { $ref: '#/components/schemas/DocumentType' }
        subjectRef: { type: string }
        createdAt: { type: string, format: date-time }
        expiresAt: { type: string, format: date-time }
        decisionIssuer:
          type: string
          enum: [MIDDLEWARE, SIGNALS_ONLY, REQUESTER_CONFIRMS]
        requiredArtifacts:
          type: array
          description: Artefactos que el sujeto debe aportar para que el flujo pueda completarse.
          items:
            type: object
            required: [slot, satisfied]
            properties:
              slot: { $ref: '#/components/schemas/ArtifactSlot' }
              satisfied: { type: boolean }
        steps:
          type: array
          items: { $ref: '#/components/schemas/StepStatus' }
    SessionState:
      type: string
      enum:
        [CREATED, COLLECTING, PROCESSING, AWAITING_SUBJECT, PENDING_REVIEW, IN_REVIEW,
         DECIDED, RETAINED, BLOCKED, FAILED, EXPIRED, ABANDONED, CANCELLED, PURGED]
    ArtifactSlot:
      type: string
      enum: [DOC_FRONT, DOC_BACK, SELFIE, LIVENESS_VIDEO, PROOF_OF_ADDRESS]
    StepStatus:
      type: object
      required: [stepId, capability, state]
      properties:
        stepId: { type: string }
        capability:
          type: string
          description: Identificador versionado de la capacidad, por ejemplo `ocr.document.v1`.
        state:
          type: string
          enum: [PENDING, READY, RUNNING, RETRYING, FALLBACK, SUCCEEDED, NEGATIVE, INCONCLUSIVE, FAILED, SKIPPED]
        provider: { type: string, nullable: true, description: Proveedor que ejecutó el paso. }
        startedAt: { type: string, format: date-time, nullable: true }
        finishedAt: { type: string, format: date-time, nullable: true }
        attempts: { type: integer, minimum: 0 }
    ArtifactUploadRequest:
      type: object
      required: [slot, contentType, sizeBytes]
      properties:
        slot: { $ref: '#/components/schemas/ArtifactSlot' }
        contentType: { type: string, example: image/jpeg }
        sizeBytes: { type: integer, minimum: 1 }
    ArtifactUploadTicket:
      type: object
      required: [artifactId, uploadUrl, method, expiresAt]
      properties:
        artifactId: { type: string }
        uploadUrl: { type: string, format: uri }
        method: { type: string, enum: [PUT, POST] }
        headers:
          type: object
          additionalProperties: { type: string }
          description: Cabeceras que deben acompañar la carga, incluidas las de cifrado del lado del servidor.
        expiresAt: { type: string, format: date-time }
    IdentityClaimSet:
      type: object
      description: |
        Conjunto canónico de datos de identidad, normalizado desde cualquier país y formato.
        Cada campo lleva su procedencia y su confianza para que la decisión sea auditable.
      properties:
        firstName: { $ref: '#/components/schemas/Claim' }
        lastName: { $ref: '#/components/schemas/Claim' }
        idNumber: { $ref: '#/components/schemas/Claim' }
        birthDate: { $ref: '#/components/schemas/Claim' }
        expiryDate: { $ref: '#/components/schemas/Claim' }
        issuingState: { $ref: '#/components/schemas/Claim' }
        nationality: { $ref: '#/components/schemas/Claim' }
        sex: { $ref: '#/components/schemas/Claim' }
    Claim:
      type: object
      required: [value, source, confidence]
      properties:
        value: { type: string }
        source:
          type: string
          enum: [MRZ, OCR, LLM, SUBJECT_DECLARED, REGISTRY, HUMAN_REVIEW]
        confidence: { type: number, minimum: 0, maximum: 1 }
        corroboratedBy:
          type: array
          items: { type: string }
          description: Otras fuentes que coinciden con este valor. Un dato corroborado por MRZ y OCR es más fuerte que uno solo del LLM.
    DecisionResult:
      type: object
      required: [sessionId, outcome, issuer, source, decidedAt, signals]
      properties:
        sessionId: { type: string }
        outcome:
          type: string
          enum: [APPROVED, REJECTED, MANUAL_REVIEW, SIGNALS_ONLY, INCONCLUSIVE]
        issuer:
          type: string
          enum: [MIDDLEWARE, SIGNALS_ONLY, REQUESTER_CONFIRMS]
        source:
          type: string
          enum: [AUTOMATED_POLICY, HUMAN_REVIEWER, REQUESTER, SYSTEM_TIMEOUT]
        decidedAt: { type: string, format: date-time }
        identity: { $ref: '#/components/schemas/IdentityClaimSet' }
        signals:
          type: array
          description: Señales que sustentan el veredicto, con su umbral aplicado. Esto es lo que hace auditable la decisión.
          items:
            type: object
            required: [name, value, threshold, passed]
            properties:
              name: { type: string, example: biometrics.facematch.similarity }
              value: { type: number }
              threshold: { type: number }
              passed: { type: boolean }
              provider: { type: string }
              modelVersion: { type: string, description: Versión del modelo que produjo la señal. Necesaria para reproducir la decisión años después. }
        reasons:
          type: array
          description: Códigos de razón estables y legibles por máquina.
          items: { type: string, example: MRZ_CHECKDIGIT_MISMATCH }
        auditChainHead: { type: string, description: Resumen del último evento de la cadena de auditoría al momento de decidir. }
    ReviewCase:
      type: object
      required: [sessionId, state, queuedAt, reasons]
      properties:
        sessionId: { type: string }
        state: { type: string, enum: [PENDING_REVIEW, IN_REVIEW] }
        queuedAt: { type: string, format: date-time }
        assignedTo: { type: string, nullable: true }
        slaExpiresAt: { type: string, format: date-time, nullable: true }
        reasons: { type: array, items: { type: string } }
...
    Problem:
      type: object
      description: Respuesta de error conforme a RFC 9457.
      required: [type, title, status, code]
      properties:
        type: { type: string, format: uri, default: about:blank }
        title: { type: string }
        status: { type: integer }
        detail: { type: string, description: "Descripción legible. Nunca contiene datos personales." }
        instance: { type: string }
        code:
          type: string
          description: Código estable de error del middleware.
          examples: [FLOW_SPEC_NOT_FOUND, TENANT_MISMATCH, CAPABILITY_NOT_AUTHORIZED, IDEMPOTENCY_CONFLICT, ARTIFACT_DIGEST_MISMATCH, PROVIDER_UNAVAILABLE]
        correlationId: { type: string }
        retryable: { type: boolean }
  responses:
    BadRequest:
      description: Petición malformada.
      content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } }
    Unauthorized:
      description: Token ausente, expirado o inválido.
      content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } }
    Forbidden:
      description: El inquilino del token no tiene acceso al recurso, o la capacidad no está autorizada para él.
      content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } }
    NotFound:
      description: |
        El recurso no existe **o pertenece a otro inquilino**. No se distingue entre ambos casos
        a propósito: revelar la existencia de un recurso ajeno ya es una fuga.
      content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } }
    Conflict:
      description: El estado actual de la sesión no admite esta operación, o la clave de idempotencia se reutilizó con un cuerpo distinto.
      content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } }
    UnprocessableEntity:
      description: La petición es sintácticamente válida pero no puede procesarse.
      content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } }
    TooManyRequests:
      description: Se superó el límite de tasa del plan de uso del inquilino.
      headers:
        Retry-After: { schema: { type: integer } }
      content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } }